requirements.txt
PySide6>=6.8.0
mss>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=8.0.0
pyobjc>=11.0 